Refactor project structure with package organization
Organize classes into structured Java packages

The previous architecture had all classes placed in the default package, which, while functional, did not fully utilize the organizational benefits of Java packages. As the codebase grows, a more structured package organization becomes crucial for maintainability and clarity.

This commit:
- Organizes all classes into a dedicated `duke` package, enhancing project structure.
- Further segregates classes into sub-packages:`duke.task` and `duke.command`. This categorization aligns with each class's functionality.
- Places task-related classes such as `Task`, `Todo`, `Deadline`, and `Event` into `duke.task`.
- Groups command processing classes like `Command`, `AddCommand`, `DeleteCommand`, etc., into `duke.command`.

The `duke.Duke` main class is also updated to reflect these changes.

By arranging classes into distinct packages, the project now benefits from improved organization, easier navigation, and better encapsulation of related functionalities.

These changes are a step towards aligning the project more closely with established Java development practices, enhancing the overall design and scalability of the application.
