+++ src/main/java/duke/command/AddCommand.java
@@ -1,3 +1,14 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.Ui;
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Storage;
+import duke.task.TaskList;
+import duke.task.Todo;
+import duke.task.DateTimeUtil;
+
 import java.io.IOException;
 import java.time.LocalDateTime;
