+++ src/main/java/duke/command/Command.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.task.Storage;
+import duke.task.TaskList;
+import duke.Ui;
+
 import java.io.IOException;
 
 public abstract class Command {
