+++ src/main/java/duke/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.task.Storage;
+import duke.task.TaskList;
+import duke.Ui;
+
 public class ExitCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
