+++ src/main/java/duke/task/Task.java
@@ -1,4 +1,6 @@
-class Task {
+package duke.task;
+
+public class Task {
     protected String description;
     protected boolean isDone;
